requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
rembg[gpu]==2.0.57
Pillow==10.3.0
